requirements.txt
contourpy==1.3.0
cycler==0.12.1
filelock==3.13.1
fonttools==4.55.0
fsspec==2024.2.0
importlib_resources==6.4.5
Jinja2==3.1.3
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==24.2
pillow==10.2.0
pynvml==11.5.3
pyparsing==3.2.0
python-dateutil==2.9.0.post0
six==1.16.0
sympy==1.13.1
torchsummary==1.5.1
typing_extensions==4.9.0
zipp==3.21.0
